Add new logic to playback overlay for ff/rewind
This adds in some nrew logic to intercept the fast forward and rewind
keys presses if the popup panel is not in view.

If the media playing is not live TV, it will skip forwards or backwards
the appropriate amount of seconds. It will show the overlay as expected]
and then iniatite the fadeout timer.