Add Playback speed controls (#1302)
* Add basic mocks to unit test PlaybackController

Adds the various basic mocks required to unit test Playback Controller.

At a high level the steps we perform are:
- Prep mocks for deps injected by constructor
- Prep mocks for Koin, then pass these as singletons
- Patch up various calls out to Android SharedPrefs ...etc.

This allows us to construct a PlaybackController in a JVM context now.
These mechanisms will be useful for later in the patch series too....

* Add setPlayback impl and tests to controller/view

Adds setPlaybackSpeed to the view and controller.
The controller has associated tests which built on the earlier ones.

Unfortunately the view also has logic in it and I've tried to (ab)use
mocks to factor out the ExoPlayer factory. However because it's a static
final the mock will always defer to the real impl.

I could introduce some sort of interface and abstract out the calls to
the underlying players using an adapter pattern so we can test this. But
alas, the player rewrite will obselete this and I'm 8+ hours into
getting tests to work...

* Add VideoSpeedController and tests

Adds a new controller for VideoSpeed, to help encapsulate items
following SRP.
This also adds a mechanism to track the users most recent speed
selection, so switching from videos / to other media and back will
restore the user's selection on the next video.

* Switch to using Enums to representing speed

This makes it much easier to compare for equality and has built in
indexing making the menu much easier to work with.

In comparison using doubles directly would have required us to add a
comparison lambda in the view to correctly detect what was selected /
determine what is selected.

* Wire up view for Playback speed controls

Wires up the view component for playback speed controls, based on the
pattern used in various other action items.

* Preserve playback speed between videos

- Update the PlaybackController to internally track the last playback
speed, but still treat it as a black box from all external callers
- Add a test we don't throw a NPE trying to set it before init
- There's no test for checking this value gets forwarded (see below)
- Change the activity to create the speed controller whilst the
ControlGlue is being created.

This is unfortunately a bit of a hack. There isn't a signal for
onPlaybackStarted or similar we can attach to from the playback
controller so we have to mess with the internals. Something I've tried
to avoid greatly.

I've tried adding a unit test with mocks, which was surprisingly going
well. Until we create LibVLC objects mid-way. These needs some sort of
factory pattern to inject mocks else it will query Android for playback
capabilities failing the test.

I'm going to leave this particular test as a manual case as I extracting
factories out with the pending rewrite is a futile effort.

* Fix various nits / warnings from CI

This includes:
- Using magic values in our enum
- Impliclt Locale issues
- Nullability checks

* Port drawable from Exoplayer into drawables

Ports the playback speed drawable into the projects list of drawable
items, following PR feedback.

The license is preserved (Apache 2.0) and compatible with the GPL
license of this project.

* Apply various code improvements to PlaybackSpeedAction

Includes the following from the PR review:
- Whitespace tidy up
- Implicitly return from lambda instead of explicitly naming it
- Use apply to remove a significant amount of boilerplate

* Add check that setPlaybackSpeed is sane

This will ignore any values that are clearly incorrect i.e. <0.25.

This means the playback controller can still have a sentinel value of
-1.0 to ensure the model controlling speed is in-place correctly. Whilst
not breaking user playback by default if something goes wrong.

As this file directly interfaces with the video manager (which holds
surfaces) we can't easily make a test for it, so leave this block
untested

* DI interface instead of concrete type

Inject the interface for shared preferences rather than the concrete
type.
This avoids Mockk trying to grab the abstract class underneath, which
would run the real impl instead. It also brings our DI in this file more
in-line with the commonly used pattern

* Address PR comments

- Revert some noisy automatic formatting
- Move speed steps out of companion object
- Drop supuerfluous interface

* Tidy up VideoSpeedController from PR review

- Use Kotlin a kotlin setter directly
- Make the variable public
- (Small amount of rewiring to make everything work)

* Drop PlaybackController tests and revert DI

Drop PlaybackController test. We can't simply inject by interface as
this isn't type safe with how it's currently implemented.

Since it's likely PlaybackController is going to undergo a major rework
with pending future work, let's simply drop the tests and revert the
associated changes.