Grab LeanbackFragment through videoAdapter
Access the CustomPlaybackOverlayFragment and
LeanbackOverlayFragment through videoAdapter. This was fundamentally
what CustomPlaybackTransportGlue did on behalf of the callers.

Overall it's a code smell that we retrieve the view through the
VideoAdapter, but without serious re-architecting it's better to keep
our call sites consistent for if/when we refactor these fragments out of
the video adapter.